Split todoFooter into _todoItem, _todoLabel
_todoItem:
  Replicates exactly old behaviour of todoFooter (parent just a
  passthrough for our stores).

_todoLabel:
  Stub file for now (return null, logs 'unimplemented' message).
  Eventually, will add "label" buttons to filter todos with assoc-
  iated labels.
