+++ src/components/todoFooter/_todoLabel.js
@@ -1,0 +1,18 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+import {observer} from 'mobx-react';
+
+export default class TodoLabel extends React.Component {
+    /*
+     * TODO Implement "label" buttons
+     *
+     * Just log and return null for now
+     *
+     */
+    render() {
+
+        console.log("TodoLabel::render() /* [ UNIMPLEMENTED ] */");
+
+        return null;
+    }
+}
